Add help-text footer to Options menu
On all options sections _except_ key bindings, add a footer that explains
the currently-selected option, based on the description of the underlying
config item.

Note that this doesn't "reserve" an area at the bottom of the screen, so it
 is possible for a long Options section to run over the footer.  However,
this is currently only a risk for the key bindings section, because none of
 the others are anywhere near long enough for this to be a problem, at
typical modern display resolutions.