Add .NET Test GitHub action
This will clone the repo, restore it, build it, and run `dotnet test`
from the root whenever there is a change made to the dev branch.

I have confirmed that this _currently_ passes in dev, and that it fails if
I intentionally add a failing statement to a unit test.