Prevent silent failure for valid commands that are not console commands
For example, "load" is currently implemented as a console command, but
"save" and "quicksave" are not.  However, there are command constants for
these.  Because of how the command lookup logic works, the user could
type "save" into the console, and it would silently fail rather than
telling them that this isn't a console command.