Fix save/load menu quirks (#630)
* Lock out Save and Quicksave during intermission or endgame
Saving when one of those layers is displayed will produce a broken save.
Reloading such a save results in the intermission/end NOT displaying, and
the player is frozen in place, unable to move or rotate.

* Prevent silent failure for valid commands that are not console commands

For example, "load" is currently implemented as a console command, but
"save" and "quicksave" are not.  However, there are command constants for
these.  Because of how the command lookup logic works, the user could
type "save" into the console, and it would silently fail rather than
telling them that this isn't a console command.

* Prevent command processor fallthrough when a menu is already displayed.
Previously, if the user tried to press one of the menu hotkeys while a
menu was already displayed, then they'd get a warning about a command with
missing args or similar.

For example, pressing F3 while the Load menu was already on the screen
would result in an error message about missing args to the "load" console
command.

* Prevent game world layer from leaking commands into console.