Route FluidSynth output through OpenAL (#728)
* Route FluidSynth over OpenAL sample streamer

Currently, FluidSynth finds its own output, which can introduce various
other problems, including changing output device at a different time than
other playing sounds, as well as using a different output path than the
rest of the game.  I suspect that this is also responsible for race
conditions that result in intermittent crashes when the application is quit
 and restarted quickly (such as when changing video color mode).

This change bridges FluidSynth's audio output over the same sample streamer
 we use for ZMusic.

* Don't do gain-halving on FluidSynth
It's not as devastatingly loud as the module player.