Improve platform-independence of window state change logic (#1164)
* Directly set window size/position/mode using GLFW
These can only be set one-at-a-time in OpenTK, which is a little silly
and introduces some weird bugs because they're unified in a single command
in GLFW.  The main issues seem to revolve around cached state in OpenTK,
because "WindowState" is not a thing that GLFW actually has, rather it's
an abstraction in OpenTK, and it's cached in the NativeWindow.

Setting all of these parameters in one shot appears to be the most reliable
way on Linux (X), and allows dependable switching between full-screen (ex)
and windowed modes.  There's still a weird thing that sometimes pops up
when switching between full-screen (borderless) and windowed.

This needs testing on Windows and Linux (Wayland).

* Remove isLinuxWayland test
I don't think we actually need this

* Handle position for borderless window mode, recovery to windowed

* Remove Linux GLFW wait states
These don't seem necessary if we're not spamming the window manager.

* Add comments clarifying why we do some of these things