Enable more .NET analyzers, fix issues from Performance analyzer (#1243)
* Enable Security, Interop, SingleFile, and Performance analyzers

* Suppress warnings related to use of MD5
While this is no longer a "secure" cryptographic algorithm, we are only
using it to compare thumbprints of WAD files and such, so the potential
for harm is limited.

* Address recommendations from Performance analyzers
These almost all fall into the following categories:
1.  Make methods static when they do not access any instance data
2.  Seal private classes that have no derived classes
3.  Do not do double-lookups on Dictionary<TKey, TValue>--use Try method
4.  Prefer concrete return types over interfaced return types when there
is only one implementation of that interface

Most of these were done automatically with the code refactoring tool in VS

* Fix a couple more warnings from the analyzer

* Make program and VAO "unbind" back into instance
methods.  While this isn't strictly correct, it
does make the code read more naturally.